Fix use of Cond.Signal to avoid deadlock
Previouldy, unlocking a key could signal any of the waiting goroutines, even if they were waiting on a different key than the one being unlocked.  This can cause a deadlock as the goroutines waiting on a particular key may never see it unlocked.  This is fixed by using a separate condition variable for each key, ensuring that only a goroutine that is waiting for that key is awakened.